requirements.txt
accelerate==1.0.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.1
argcomplete==3.5.1
async-timeout==4.0.3
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
datamodel-code-generator==0.26.1
datasets==3.0.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
gekko==1.2.1
genson==1.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse>=0.4.0
huggingface-hub==0.25.2
humanfriendly==10.0
identify==2.6.1
idna==3.10
inflect==5.6.2
iniconfig==2.0.0
instructor==1.3.7
isort==5.13.2
Jinja2==3.1.4
jiter==0.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
networkx==3.4.1
nodeenv==1.9.1
numpy==1.26.4
openai>=2.26.0,<3.0.0
optimum==1.23.1
ormsgpack>=1.12.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pydantic>=2.7.4
pydantic_core>=2.23.4
pydantic-settings>=2.10.1
Pygments==2.18.0
pynvml==11.5.3
PyQRCode==1.2.1
pyreadline3==3.5.4
pytest==8.3.3
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rouge==1.0.1
rpds-py==0.20.0
safetensors==0.4.5
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3

uuid-utils>=0.12.0
virtualenv==20.26.6
xxhash>=3.5.0
yarl==1.9.11
